Increased default cycle count to emulate to 8 million.
Some XMBurner tests needed it, and below about 32M instructions it should
be fine (the goal is that the emulator normally should return after about
a second of work on an average host, preventing runaway tests taking too
long).

Also fixed the related documentation (which described it as instruction
count which isn't correct).